Makes the ortac mode be plugins
This commit bring the plugin architecture to ortac. The idea is that
each plugin exposes a `Cmdliner.Cmd` that the ortac executable will
gather and group as subcommands.

We try to bring the plugin architecture to ortac with minimal change in
the repository architecture.

NOTE: in order to comply with the `(implicit_transitive_deps false)`
stanza of the dune-project, it is necessary to declare a dependency
to a dune private library.