Map the full Gospelstdlib to the Ortac_runtime.Gospelstdlib module
Instead of hard-coding an association table that maps Gospel Stdlib
values to their equivalent OCaml code, expect the target runtime to
provide its own implementation in the Ortac_runtime.Gospelstdlib module,
following exactly the same structure as the Gospel Stdlib.
Consequently, Ortac_runtime_intf.Gospelstdlib could be generated (using
fold_namespace for instance...). It is still handwritten, because the
implementation is real only partial for now.

That restructuring makes it clearer which parts of the runtime is
exposed to support the specification translations and which parts should
really only be used by the generator (`Z.forall`, `Z.exists`). On the
other hand, the generator does call the functions that are provided by
the `Gospelstdlib` sub-module (`pred`, etc.)

To ensure that `Ortac_runtime.Z` does not shadow `Z` in the
implementation of the `Gospelstdlib`, `Gospelstdlib` is implemented
before

The expected test results are updated accordingly