Add a config and init function for it
The config will hold the information necessary to build the qcheck-stm
test and that are given by the user such as the type of the system under
test and the init function.

The Config.init function check that there is actually a type for the
given sut in the given module. It also initialize the translation
context from Ortac_core.