Support `checks` clauses
Some refactoring along the way as we reuse some code.
Build a conditional expression iff the list of translated `checks`
clauses is not empty.
If all of the `checks` clauses are true, then we fall back on the `rhs`
built for `xposts` previous verification (exceptional postconditions).
Otherwise, result should be `Error (Invalid_argument _)`.
Update tests.