Implement the Bag and Set modules of the Gospel Stdlib
Define BagSet to factorise the implementations of Bag and Set

`bag` and `set` are implemented as decreasing sorted lists where each
element appears at most once, using the polymorphic comparison.

For `bag`s, the list contains pairs, where `snd x` is the number of
occurrences of `fst x`, with `snd x > 0`.

Rationale: obviously, the time complexity of these implementations is
not ideal, but:
- it is a purely functional implementation (so it can be used with
  QCheck-STM),
- it is compatible with the polymorphic equality: Ortac uses OCaml's
  `(=)` to implement Gospel's `(=)`, this ensures that `set`s and `bag`s
  can be tested for equality in the generated code,
- they are sorted in decreasing order so that the resulting order
  (according to the polymorphic comparison) on sets and bags is more
  natural.