Make postcond return the reason of failure rather than a bool
The Spec module contains a dummy postcond function, always returning
true. The new postcond function return an `Ortac_runtime.report option`.
Checking one term return a value of such a type, with a singleton for
the list. Results of checking several terms will be combine using an
append function defined in the runtime.