Make `ortac_postcond` also return the expected returned value
Also adapt the runtime so that the commit build.
When a returned value can't be computed based on the information
contained in the Gospel specification, a warning is displayed to the
user at generation and a dummy one is placed to inform the message
printer that the expected returned value is unknown.