Remove types tests for wrapper
These tests ask to access OCaml fields in Gospel specification. This
should soon become illegal Gospel. Also, the symbol for field access is
not the same as the one for field definition in the Gospel ast (not the
same tags), which will lead to trouble when fixing how to print the
field access the `Ocaml_of_gospel.term`