Add ability to run tests in separate processes
When a test causes a segmentation fault, ortac/qcheck-stm
can not report the (simplified) trace which lead to that
fault. This is due to the fact, that the test runs in the
same process as the ortac runtime. In order to get more
output, the runtime is adapted to allow forking tests in a
separate process.

This also solves problems regarding non-terminating
programs.