Make ortac qcheck-stm generate STM_domain
This first version makes the following design choice: when testing in a
parallel context (with the `domain` flag given on the cli), we don't
store the new SUTs created during the tests.

This weakens how the SUT-returning functions are tested as we will
only be able to distinguish between normal and exceptional behaviour and
not inspect the newly created SUTs in the rest of the generated program.

The rational is about ease of implementation: we would have to
differentiate between the commands from the sequential prefix and those
from one of the parallel tails. The idea is then to make everything
works without this feature and add it later.