Make ortac qcheck-stm generate STM_domain
This first version makes the following design choice: when testing in a
concurrent context, we don't store the new SUTs created during the
tests.

This weakens the tests of the sequential prefix compared to the
sequential tests.

The rational is about ease of implementation, allowing for tesing with
domains without having to differentiate between commands ran in the
sequential prefix and those ran in one of the spawned domain.