[QCheck-STM] Keep frequencies locations
Specializes FrequenciesMap to handle locations transparently for the
client (`smt_of_ir`) and adapt the latter.

Location are the one of the value's name which is used as a key in the
map. But we store them in the map's values as they would mess the
comparison function when searching for a binding.