[QCheck-STM] Use `arb_triple` in domain runtime
`arb_triple` allows to differentiate between the generator used in each
part of the generated triplet.

This commit also extends the module signature that the functors expect
so that the newly generated `arb_cmd_*` are available.

The behaviour of the generated code is expected to be the same as
before.